cleanup collector_probe.c
